refactor: rename and move `CustomImg`
